+++ packages/core/src/global/config/config.ts
@@ -1,13 +1,13 @@
-import { DsLogger } from '@utils'
-import { DS_ANIMATION_KEY } from './config.const'
+import { BalLogger } from '../log'
+import { BALOISE_ANIMATION_KEY } from './config.const'
 import { defaultConfig } from './config.default'
-import { DsBrand, DsConfig, DsConfigState, DsIcons, DsLanguage, DsRegion } from './config.types'
-import { DsConfigObserver } from './observable/observer'
+import { BalBrand, BalConfig, BalConfigState, BalIcons, BalLanguage, BalRegion } from './config.types'
+import { BalConfigObserver } from './observable/observer'
 
 export class Config {
-  private _componentObservers: DsConfigObserver[] = []
-  private _observers: DsConfigObserver[] = []
-  private _config: DsConfigState = defaultConfig
+  private _componentObservers: BalConfigObserver[] = []
+  private _observers: BalConfigObserver[] = []
+  private _config: BalConfigState = defaultConfig
 
   public _jmp?: (c: any) => any
   public _raf?: (c: any) => number
@@ -15,37 +15,37 @@ export class Config {
   public _rel?: (el: any, eventName: string, listener: any, options: any) => void
   public _ce?: (eventName: string, opts?: any) => any
 
-  get brand(): DsBrand {
-    return this._config.brand
-  }
-
-  set brand(brand: DsBrand) {
-    if (brand !== this._config.brand) {
-      this._config.brand = brand
-      this._notify()
-    }
-  }
-
   get locale(): string {
     return `${this._config.language}-${this._config.region}`
   }
 
-  get region(): DsRegion {
+  get region(): BalRegion {
     return this._config.region
   }
 
-  set region(region: DsRegion) {
+  set region(region: BalRegion) {
     if (region !== this._config.region) {
       this._config.region = region
       this._notify()
     }
   }
 
-  get language(): DsLanguage {
+  get language(): BalLanguage {
     return this._config.language
   }
 
-  set language(language: DsLanguage) {
+  get brand(): BalBrand {
+    return this._config.brand
+  }
+
+  set brand(brand: BalBrand) {
+    if (brand !== this._config.brand) {
+      this._config.brand = brand
+      this._notify()
+    }
+  }
+
+  set language(language: BalLanguage) {
     if (language !== this._config.language) {
       if (this._config.allowedLanguages.includes(language)) {
         this._config.language = language
@@ -57,22 +57,22 @@ export class Config {
     }
   }
 
-  get allowedLanguages(): DsLanguage[] {
+  get allowedLanguages(): BalLanguage[] {
     return this._config.allowedLanguages
   }
 
-  set allowedLanguages(allowedLanguages: DsLanguage[]) {
+  set allowedLanguages(allowedLanguages: BalLanguage[]) {
     if (allowedLanguages !== this._config.allowedLanguages) {
       this._config.allowedLanguages = allowedLanguages
       this._notify()
     }
   }
 
-  get icons(): DsIcons {
+  get icons(): BalIcons {
     return this._config.icons
   }
 
-  set icons(icons: DsIcons) {
+  set icons(icons: BalIcons) {
     this._config.icons = {
       ...this._config.icons,
       ...icons,
@@ -80,11 +80,11 @@ export class Config {
     this._notify()
   }
 
-  get logger(): DsLogger {
+  get logger(): BalLogger {
     return this._config.logger
   }
 
-  set logger(logger: DsLogger) {
+  set logger(logger: BalLogger) {
     this._config.logger = { ...logger }
     this._notify()
   }
@@ -107,7 +107,7 @@ export class Config {
     this._notify()
   }
 
-  attach(observer: DsConfigObserver): void {
+  attach(observer: BalConfigObserver): void {
     const isExist = this._observers.includes(observer)
     if (isExist) {
       return console.log('Subject: Observer has been attached already.')
@@ -117,7 +117,7 @@ export class Config {
     observer.configChanged(this._config)
   }
 
-  detach(observer: DsConfigObserver): void {
+  detach(observer: BalConfigObserver): void {
     const observerIndex = this._observers.indexOf(observer)
     if (observerIndex === -1) {
       return console.log('Subject: Nonexistent observer.')
@@ -126,7 +126,7 @@ export class Config {
     this._observers.splice(observerIndex, 1)
   }
 
-  attachComponent(observer: DsConfigObserver): void {
+  attachComponent(observer: BalConfigObserver): void {
     const isExist = this._componentObservers.includes(observer)
     if (isExist) {
       return console.log('Subject: Observer has been attached already.')
@@ -136,7 +136,7 @@ export class Config {
     observer.configChanged(this._config)
   }
 
-  detachComponent(observer: DsConfigObserver): void {
+  detachComponent(observer: BalConfigObserver): void {
     const observerIndex = this._componentObservers.indexOf(observer)
     if (observerIndex === -1) {
       return console.log('Subject: Nonexistent observer.')
@@ -149,7 +149,7 @@ export class Config {
     return JSON.stringify(this._config)
   }
 
-  reset(config: DsConfig) {
+  reset(config: BalConfig) {
     this._config = {
       ...this._config,
       ...config,
@@ -178,7 +178,7 @@ export const config = /*@__PURE__*/ new Config()
 
 export const configFromLocalStorage = (win: Window): any => {
   try {
-    const animated = JSON.parse(win.localStorage.getItem(DS_ANIMATION_KEY) || 'true')
+    const animated = JSON.parse(win.localStorage.getItem(BALOISE_ANIMATION_KEY))
 
     if (animated !== null) {
       return {
@@ -187,7 +187,7 @@ export const configFromLocalStorage = (win: Window): any => {
     }
 
     return {}
-  } catch {
+  } catch (e) {
     return {}
   }
 }